+++ app/evaluation_tests.py
@@ -23,13 +23,13 @@ class TestEvaluationFunction():
     """
 
     # Import tests that makes sure that mathematical expression comparison works as expected
-    from .symbolic_comparison_evaluation_tests import TestEvaluationFunction as TestSymbolicComparison
+    from .tests.symbolic_comparison_evaluation_tests import TestEvaluationFunction as TestSymbolicComparison
 
     # Import tests that makes sure that physical quantities are handled as expected
-    from .quantity_comparison_evaluation_tests import TestEvaluationFunction as TestQuantities
+    from .tests.quantity_comparison_evaluation_tests import TestEvaluationFunction as TestQuantities
 
     # Import tests that corresponds to examples in documentation and examples module
-    from .example_tests import TestEvaluationFunction as TestExamples
+    from .tests.example_tests import TestEvaluationFunction as TestExamples
 
     def test_eval_function_can_handle_latex_input(self):
         response = r"\sin x + x^{7}"
