+++ app/quantity_comparison_preview.py
@@ -1,11 +1,11 @@
-from .expression_utilities import (
+from .utility.expression_utilities import (
     find_matching_parenthesis,
     parse_expression,
     SymbolDict,
     sympy_to_latex,
 )
 
-from .preview_utilities import (
+from .utility.preview_utilities import (
     Params,
     Preview,
     Result,
@@ -13,8 +13,8 @@ from .preview_utilities import (
     sanitise_latex,
 )
 
-from .physical_quantity_utilities import SLR_quantity_parser as quantity_parser
-from .physical_quantity_utilities import SLR_quantity_parsing as quantity_parsing
+from .utility.physical_quantity_utilities import SLR_quantity_parser as quantity_parser
+from .utility.physical_quantity_utilities import SLR_quantity_parsing as quantity_parsing
 
 
 def fix_exponents(response):
