+++ app/tests/quantity_comparison_evaluation_tests.py
@@ -3,8 +3,8 @@ import os
 
 # Import necessary data and reference cases for tests
 from .slr_quantity_tests import slr_strict_si_syntax_test_cases, slr_natural_si_syntax_test_cases
-from .evaluation import evaluation_function
-from .unit_system_conversions import (
+from ..evaluation import evaluation_function
+from ..utility.unit_system_conversions import (
     set_of_SI_prefixes,
     set_of_SI_base_unit_dimensions,
     set_of_derived_SI_units_in_SI_base_units,
