+++ app/tests/symbolic_comparison_preview_tests.py
@@ -1,8 +1,8 @@
 import os
 import pytest
 
-from .preview_utilities import Params, extract_latex
-from .symbolic_comparison_preview import preview_function
+from ..utility.preview_utilities import Params, extract_latex
+from ..symbolic_comparison_preview import preview_function
 from .symbolic_comparison_evaluation_tests import elementary_function_test_cases
