+++ app/utility/criteria_graph_utilities.py
